fix(economy): keep projection leftover cases out of support LOC ratchet (#4550)
Move the large projectProvinceExtraction pins into the economy test tree so
colonizethis_economy_test_support stays under the 6590 aggregate ceiling.
